Stopped using deprecated aliases of builtin types. This is required to avoid warnings starting with NumPy 1.20.0.

+++ lib/iris/tests/test_cell.py
@@ -130,7 +130,7 @@ class TestCells(tests.IrisTest):
         self.assertEqual(self.d.__ne__(Terry()), NotImplemented)
 
     def test_numpy_int_equality(self):
-        dtypes = (np.int, np.int16, np.int32, np.int64)
+        dtypes = (int, np.int16, np.int32, np.int64)
         for dtype in dtypes:
             val = dtype(3)
             cell = iris.coords.Cell(val, None)
@@ -138,7 +138,7 @@ class TestCells(tests.IrisTest):
 
     def test_numpy_float_equality(self):
         dtypes = (
-            np.float,
+            float,
             np.float16,
             np.float32,
             np.float64,
